requirements.txt
Django==5.2.7
django-ckeditor==6.7.1
Pillow==10.4.0
